Separate collect from hoist
